@@ -26,27 +26,34 @@
     "node": ">=24.12.0"
   },
   "dependencies": {
-    "@angular/animations": "^21.0.6",
-    "@angular/cdk": "^21.0.3",
-    "@angular/common": "^21.0.0",
-    "@angular/compiler": "^21.0.0",
-    "@angular/core": "^21.0.0",
-    "@angular/forms": "^21.0.0",
-    "@angular/platform-browser": "^21.0.0",
-    "@angular/router": "^21.0.0",
-    "@angular/service-worker": "^21.0.0",
-    "@primeng/themes": "^21.0.2",
+    "@angular/animations": "^21.1.1",
+    "@angular/cdk": "^21.1.1",
+    "@angular/common": "^21.1.1",
+    "@angular/compiler": "^21.1.1",
+    "@angular/core": "^21.1.1",
+    "@angular/forms": "^21.1.1",
+    "@angular/platform-browser": "^21.1.1",
+    "@angular/router": "^21.1.1",
+    "@angular/service-worker": "^21.1.1",
+    "@jsverse/transloco": "^8.2.1",
+    "@jsverse/transloco-keys-manager": "^8.1.0",
+    "@primeuix/themes": "^2.0.3",
+    "@tailwindcss/postcss": "^4.1.18",
+    "postcss": "^8.5.6",
     "primeflex": "^4.0.0",
     "primeicons": "^7.0.0",
-    "primeng": "^21.0.2",
+    "primeng": "^21.0.4",
     "rxjs": "~7.8.0",
+    "tailwindcss": "^4.1.18",
+    "tailwindcss-primeui": "^0.6.1",
     "tslib": "^2.3.0"
   },
   "devDependencies": {
-    "@angular/build": "^21.0.3",
-    "@angular/cli": "^21.0.3",
-    "@angular/compiler-cli": "^21.0.0",
-    "jsdom": "^27.1.0",
+    "@angular/build": "^21.1.1",
+    "@angular/cli": "^21.1.1",
+    "@angular/compiler-cli": "^21.1.1",
+    "jsdom": "^27.4.0",
+    "sass": "^1.97.3",
     "typescript": "~5.9.2",
     "vitest": "^4.0.8"
   }
