@@ -1,34 +1,40 @@
 import { Injectable, inject } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable, firstValueFrom } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { Observable } from 'rxjs';
 
-export interface CheckAvailabilityResponse {
-  nameAvailable: boolean;
-  slugAvailable: boolean;
+export interface RegisterOrganizationRequest {
+  organizationName: string;
+  userEmail: string;
+  userFirstName: string;
+  userLastName: string;
+  userPassword: string;
 }
 
-export interface CreateOrganizationRequest {
-  name: string;
-  slug: string;
-}
-
-export interface CreateOrganizationResponse {
+export interface RegisterOrganizationResponse {
   organizationId: string;
 }
 
-@Injectable({ providedIn: 'root' })
+export interface CheckAvailabilityResponse {
+  nameAvailable: boolean;
+}
+
+@Injectable({
+  providedIn: 'root',
+})
 export class OrganizationService {
   private readonly http = inject(HttpClient);
   private readonly baseUrl = '/api/organizations';
 
-  checkSlug(slug: string): Observable<boolean> {
-    return this.http
-      .get<CheckAvailabilityResponse>(`${this.baseUrl}/check-availability?slug=${slug}`)
-      .pipe(map((res) => res.slugAvailable));
+  registerOrganization(
+    request: RegisterOrganizationRequest,
+  ): Observable<RegisterOrganizationResponse> {
+    return this.http.post<RegisterOrganizationResponse>(this.baseUrl, request);
   }
 
-  async create(req: CreateOrganizationRequest): Promise<CreateOrganizationResponse> {
-    return firstValueFrom(this.http.post<CreateOrganizationResponse>(this.baseUrl, req));
+  checkAvailability(name: string): Observable<CheckAvailabilityResponse> {
+    const params = new HttpParams().set('name', name);
+    return this.http.get<CheckAvailabilityResponse>(`${this.baseUrl}/check-availability`, {
+      params,
+    });
   }
 }
