@@ -1,6 +1,31 @@
-using ProperTea.ServiceDefaults.ErrorHandling;
+using ProperTea.Infrastructure.Common.ErrorHandling;
 
 namespace ProperTea.Landlord.Bff.Organizations;
+
+public class OrganizationClientAnonymous(HttpClient httpClient)
+{
+    public async Task<CheckAvailabilityResponse> CheckAvailabilityAsync(
+        string? name = null,
+        CancellationToken ct = default)
+    {
+        var query = new List<string>();
+        if (!string.IsNullOrWhiteSpace(name))
+            query.Add($"name={Uri.EscapeDataString(name)}");
+
+        var queryString = query.Count > 0 ? "?" + string.Join("&", query) : "";
+        return (await httpClient.GetFromJsonAsync<CheckAvailabilityResponse>(
+            $"/organizations/check-availability{queryString}", ct))!;
+    }
+
+    public async Task<RegisterOrganizationResponse> RegisterOrganizationAsync(
+        RegisterOrganizationRequest request,
+        CancellationToken ct = default)
+    {
+        var response = await httpClient.PostAsJsonAsync("/organizations", request, ct);
+        await response.EnsureDownstreamSuccessAsync(ct: ct);
+        return (await response.Content.ReadFromJsonAsync<RegisterOrganizationResponse>(ct))!;
+    }
+}
 
 public class OrganizationClient(HttpClient httpClient)
 {
@@ -14,31 +39,6 @@ public class OrganizationClient(HttpClient httpClient)
         {
             return null;
         }
-    }
-
-    public async Task<RegisterOrganizationResponse> RegisterOrganizationAsync(
-        RegisterOrganizationRequest request,
-        CancellationToken ct = default)
-    {
-        var response = await httpClient.PostAsJsonAsync("/organizations", request, ct);
-        await response.EnsureDownstreamSuccessAsync(ct: ct);
-        return (await response.Content.ReadFromJsonAsync<RegisterOrganizationResponse>(ct))!;
-    }
-
-    public async Task<CheckAvailabilityResponse> CheckAvailabilityAsync(
-        string? name = null,
-        string? slug = null,
-        CancellationToken ct = default)
-    {
-        var query = new List<string>();
-        if (!string.IsNullOrWhiteSpace(name))
-            query.Add($"name={Uri.EscapeDataString(name)}");
-        if (!string.IsNullOrWhiteSpace(slug))
-            query.Add($"slug={Uri.EscapeDataString(slug)}");
-
-        var queryString = query.Count > 0 ? "?" + string.Join("&", query) : "";
-        return (await httpClient.GetFromJsonAsync<CheckAvailabilityResponse>(
-            $"/organizations/check-availability{queryString}", ct))!;
     }
 
     public async Task<AuditLogResponse?> GetAuditLogAsync(Guid id, CancellationToken ct = default)
