@@ -1,17 +1,26 @@
 import { Routes } from '@angular/router';
+import { unauthenticatedOnlyGuard } from './auth/guards/unauthenticated-only.guard';
+import { authenticatedGuard } from './auth/guards/authenticated.guard';
+import { DashboardComponent } from './features/dashboard/pages/dashboard.component';
 
 export const routes: Routes = [
-  { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
   {
-    path: 'auth',
-    loadChildren: () => import('./features/auth/auth.routes').then(m => m.AUTH_ROUTES)
+    path: '',
+    loadChildren: () => import('./pages/public.routes').then(m => m.PUBLIC_ROUTES),
+    canActivate: [unauthenticatedOnlyGuard]
+  },
+  {
+    path: 'organizations/register',
+    loadChildren: () => import('./features/organizations/organizations.routes').then(m => m.ORGANIZATION_ROUTES),
+    canActivate: [unauthenticatedOnlyGuard]
   },
   {
     path: 'dashboard',
-    loadChildren: () => import('./features/dashboard/dashboard.routes').then(m => m.DASHBOARD_ROUTES)
+    component: DashboardComponent,
+    canActivate: [authenticatedGuard]
   },
   {
-    path: 'organizations',
-    loadChildren: () => import('./features/organizations/organizations.routes').then(m => m.ORGANIZATION_ROUTES)
+    path: '**',
+    redirectTo: 'dashboard'
   }
 ];
