@@ -1,3 +1,4 @@
 // Core services barrel export
-export * from '../features/auth/services/auth.service';
+export * from '../auth/services/auth.service';
 export * from './services/config.service';
+export * from './services/search.service';
