@@ -1,8 +1,9 @@
 import { Routes } from '@angular/router';
+import { DashboardComponent } from './pages/dashboard.component';
 
 export const DASHBOARD_ROUTES: Routes = [
   {
     path: '',
-    loadComponent: () => import('./dashboard.component').then(m => m.DashboardComponent)
+    component: DashboardComponent
   }
 ];
