@@ -3,20 +3,27 @@ import { RouterLink, RouterLinkActive } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { ButtonModule } from 'primeng/button';
 import { DrawerModule } from 'primeng/drawer';
-import { AvatarModule } from 'primeng/avatar';
-import { TooltipModule } from 'primeng/tooltip';
+import { TranslocoModule } from '@jsverse/transloco';
 import { AuthService } from '@core';
+import { HeaderComponent } from './header/header.component';
 
 @Component({
   selector: 'app-layout',
   standalone: true,
-  imports: [CommonModule, RouterLink, RouterLinkActive, ButtonModule, DrawerModule, AvatarModule, TooltipModule],
+  imports: [CommonModule, RouterLink, RouterLinkActive, ButtonModule, DrawerModule, HeaderComponent, TranslocoModule],
   templateUrl: './layout.component.html',
   styleUrls: ['./layout.component.scss']
 })
 export class LayoutComponent {
   protected readonly authService = inject(AuthService);
-  sidebarVisible = signal(true);
+  mobileMenuVisible = signal(false);
+  currentYear = new Date().getFullYear();
+
+  readonly navigationItems = [
+    { route: '/dashboard', icon: 'pi pi-home', labelKey: 'nav.dashboard' },
+    { route: '/properties', icon: 'pi pi-building', labelKey: 'nav.properties' },
+    { route: '/tenants', icon: 'pi pi-users', labelKey: 'nav.tenants' }
+  ];
 
   readonly systemHealth = resource({
     loader: async () => {
@@ -30,4 +37,8 @@ export class LayoutComponent {
   });
 
   readonly healthStatus = computed(() => this.systemHealth.value() ?? 'Checking...');
+
+  toggleMenu(): void {
+    this.mobileMenuVisible.update(v => !v);
+  }
 }
