@@ -12,18 +12,25 @@ public static class OrganizationConfiguration
         IConfiguration configuration,
         IHostEnvironment environment)
     {
+        // Load ServiceAccount once as singleton (it's just config)
+        _ = services.AddSingleton(sp =>
+        {
+            var config = sp.GetRequiredService<IConfiguration>();
+            var serviceAccountJwtPath = config["Zitadel:ServiceAccountJwtPath"]
+                ?? throw new InvalidOperationException("Zitadel:ServiceAccountJwtPath not configured");
+
+            return ServiceAccount.LoadFromJsonFile(serviceAccountJwtPath);
+        });
+
+        // Register client as singleton - Zitadel SDK should handle token refresh internally
         _ = services.AddSingleton<IExternalOrganizationClient>(sp =>
             {
                 var config = sp.GetRequiredService<IConfiguration>();
                 var logger = sp.GetRequiredService<ILogger<ZitadelOrganizationClient>>();
+                var serviceAccount = sp.GetRequiredService<ServiceAccount>();
 
                 var apiUrl = config["OIDC:Authority"]
                     ?? throw new InvalidOperationException("OIDC:Authority not configured");
-
-                var serviceAccountJwtPath = config["Zitadel:ServiceAccountJwtPath"]
-                    ?? throw new InvalidOperationException("Zitadel:ServiceAccountJwtPath not configured");
-
-                var serviceAccount = ServiceAccount.LoadFromJsonFile(serviceAccountJwtPath);
 
                 var allowInsecure = environment.IsDevelopment();
 
